do not crash on rm_rf when Dir.mktmpdir was interrupted and did not return a dir
```
ruby-2.5.3/lib/ruby/2.5.0/fileutils.rb:1456:in `path': ErrorNotifier caught exception: no implicit conversion of nil into String. Use ErrorNotifier.expects(:notify) to silence in tests (RuntimeError)
  from ruby-2.5.3/lib/ruby/2.5.0/fileutils.rb:572:in `rm_rf'
  from app/models/job_execution.rb:267:in `ensure in make_tempdir'
```

+++ app/models/job_execution.rb
@@ -264,7 +264,9 @@ class JobExecution
     dir = Dir.mktmpdir("samson-#{@job.project.permalink}-#{@job.id}")
     yield dir
   ensure
-    FileUtils.rm_rf dir # mktmpdir with block often raised errors when tempfolder got removed early
-    @repository.prune_worktree
+    if dir
+      FileUtils.rm_rf dir # mktmpdir with block often raised errors when tempfolder got removed early
+      @repository.prune_worktree
+    end
   end
 end
